requirements/requirements.txt
Django==2.0.5
Pygments==2.2.0
backcall==0.1.0
decorator==4.3.0
django-debug-toolbar==1.9.1
django-extensions==2.0.7
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
numpy==1.14.2
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pytz==2018.4
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
typing==3.6.4
wcwidth==0.1.7
